app/requirements.txt
Flask==2.2.1
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.9
Werkzeug==2.2.1
python-dotenv
